+++ buildchain/webpack-settings/app.settings.js
@@ -2,21 +2,23 @@
 
 // node modules
 require('dotenv').config();
+const path = require('path');
 
 // settings
 module.exports = {
     alias: {
+        '@': path.resolve('../src/assetbundles/imageoptimize/src'),
     },
     copyright: '©2020 nystudio107.com',
     entry: {
-        'imageoptimize': '../src/assetbundles/imageoptimize/src/js/ImageOptimize.js',
-        'field': '../src/assetbundles/imageoptimize/src/js/OptimizedImagesField.js',
-        'welcome': '../src/assetbundles/imageoptimize/src/js/Welcome.js',
+        'imageoptimize': '@/js/ImageOptimize.js',
+        'field': '@/js/OptimizedImagesField.js',
+        'welcome': '@/js/Welcome.js',
     },
     extensions: ['.ts', '.js', '.vue', '.json'],
     name: 'imageoptimize',
     paths: {
-        dist: '../../src/assetbundles/imageoptimize/dist/',
+        dist: path.resolve('../src/assetbundles/imageoptimize/dist/'),
     },
     urls: {
         publicPath: () => process.env.PUBLIC_PATH || '',