+++ src/assetbundles/imageoptimize/src/js/ImageOptimize.js
@@ -1,3 +1,3 @@
 // Import our CSS
-import '../css/app.pcss';
-import '../img/ImageOptimize-icon.svg';
+import '@/css/app.pcss';
+import '@/img/ImageOptimize-icon.svg';