+++ src/assetbundles/imageoptimize/src/js/Welcome.js
@@ -1,4 +1,4 @@
-import Confetti from '../vue/Confetti.vue';
+import Confetti from '@/vue/Confetti.vue';
 
 const vm = new Vue({
     el: "#cp-nav-content",