+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # ImageOptimize Changelog
 
+## 1.6.27 - 2021.03.28
+### Changed
+* Move settings from the `composer.json` “extra” to the plugin main class
+* Move the manifest service registration to the constructor
+* Require `nystudio107/craft-imageoptimize-imgix` version `^1.2.1`
+
+### Fixed
+* Return `null` if no transform needs to be done, so other plugins can hook in
+
 ## 1.6.26 - 2021.02.26
 ### Fixed
 * Fixed a regression that caused the image sizes to show up as `???` in the Asset Detail view